requirements.txt
absolufy-imports==0.3.1
accelerate==1.10.1
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
attrs==25.3.0
Authlib==1.6.4
beautifulsoup4==4.13.5
blinker==1.9.0
Bottleneck==1.6.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dill==0.4.0
docling==2.53.0
docling-core==2.48.1
docling-ibm-models==3.9.1
docling-parse==4.5.0
docstring_parser==0.17.0
easyocr==1.7.2
et_xmlfile==2.0.0
Faker==37.8.0
fastapi==0.116.2
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.14.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.115.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-language==2.17.2
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.38.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-google-vertexai==2.1.2
langchain-text-splitters==0.3.11
langsmith==0.4.29
latex2mathml==3.78.1
lazy_loader==0.4
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mcp==1.14.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
narwhals==2.5.0
networkx==3.5
ninja==1.13.0
numexpr==2.12.1
numpy==2.3.3
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
polyfactory==2.22.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.4
pyparsing==3.2.4
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
streamlit==1.49.1
streamlit-option-menu==0.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.9.9
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
uvloop==0.21.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
xlsxwriter==3.2.9
zipp==3.23.0
zstandard==0.25.0
